HIPCMS-666-Added-Check-For-NullReferenceException-Due-To-Resharper
Resharper check in TFS fails and hence build is getting partially
succeeded. I don't know why this error suddenly came so wierd which did
not come up in the previous commits. So I added the check. Please check
the details in the commit for the reason to add the Exception check.