Iss hipcms 666 implement role check using authv3 (#19)
* HIPCMS-666-Initial-draft-of-Authv3-in-Featuretoggle

* HIPCMS-666-Implemented-Admin-Role-Check

* HIPCMS-666-Changesinauth

* HIPCMS-666-Authv3-Reimplemented

* HIPCMS-666-Changes-in-docker-file

* HIPCMS-666-PR-Changes

* HIPCMS-666-Added-A-Comment-Just-To-Trigger-A-New-Build

* HIPCMS-666-Added-Check-For-NullReferenceException-Due-To-Resharper

Resharper check in TFS fails and hence build is getting partially
succeeded. I don't know why this error suddenly came so wierd which did
not come up in the previous commits. So I added the check. Please check
the details in the commit for the reason to add the Exception check.

* Added check for null expression

* Removed obsolete NullReferenceCheck